implemented CalculateSalesSummary

+++ SingluarSystems.Models/ProductSaleModel.cs
@@ -10,7 +10,7 @@ namespace SingluarSystems.Models
     {
         public int SaleId { get; set; }
         public int ProductId { get; set; }
-        public double SalePrice { get; set; }
+        public decimal SalePrice { get; set; }
         public int SaleQty { get; set; }
         public DateTime? SaleDate { get; set; }
     }
